# docs/swagger.yml
openapi: 3.0.1
info:
  title: core-api
  version: "0.0"
paths:
  /api/v1/user:
    get:
      summary: Search for users to add
      operationId: findAllPublic
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      responses:
        default:
          description: findAllPublic default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublicRes'
    post:
      operationId: postUser
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserReq'
        required: true
      responses:
        default:
          description: postUser default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/user/me:
    get:
      operationId: getMe
      parameters: []
      responses:
        default:
          description: getMe default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/baseJoin/me:
    get:
      description: Not Implemented Yet
      operationId: getBases
      parameters: []
      responses:
        default:
          description: getBases default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BaseJoin'
  /api/v1/baseJoin:
    post:
      description: Not Implemented Yet
      operationId: requestIntoBase
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseJoinReq'
        required: true
      responses:
        default:
          description: requestIntoBase default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseJoin'
  /api/v1/baseJoin/{baseId}:
    delete:
      description: Not Implemented Yet
      operationId: deleteBaseInvite
      parameters:
      - name: baseId
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: deleteBaseInvite default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/v1/base:
    get:
      operationId: getOwnBases
      parameters: []
      responses:
        default:
          description: getOwnBases default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRes_Base_'
    post:
      operationId: postBase
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseReq'
        required: true
      responses:
        default:
          description: postBase default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
  /api/v1/base/{baseId}:
    get:
      operationId: getBaseIfOwn
      parameters:
      - name: baseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        default:
          description: getBaseIfOwn default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
    delete:
      description: Not Implemented Yet
      operationId: makeBaseInactive
      parameters:
      - name: baseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        default:
          description: makeBaseInactive default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
    patch:
      operationId: patchBase
      parameters:
      - name: baseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseReq'
        required: true
      responses:
        default:
          description: patchBase default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
  /api/v1/base/{baseId}/members:
    delete:
      description: Not Implemented Yet
      operationId: removeMembers
      parameters:
      - name: baseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListReq_UUID_'
        required: true
      responses:
        default:
          description: removeMembers default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
  /api/v1/mission/me:
    get:
      description: Not Implemented Yet
      operationId: getOwnMissions
      parameters: []
      responses:
        default:
          description: getOwnMissions default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mission'
  /api/v1/mission/progress:
    post:
      description: Not Implemented Yet
      operationId: progressMission
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                progressReq:
                  $ref: '#/components/schemas/ProgressReq'
        required: true
      responses:
        default:
          description: progressMission default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
components:
  schemas:
    UserPublicRes:
      type: object
      properties:
        score:
          type: number
          format: double
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
    TouchBasePostgresDomain:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        creator:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      allOf:
      - $ref: '#/components/schemas/TouchBasePostgresDomain'
      - type: object
        properties:
          authKey:
            type: string
          username:
            type: string
          email:
            type: string
          score:
            type: number
            format: double
          firstName:
            type: string
          lastName:
            type: string
          imageUrl:
            type: string
    UserReq:
      type: object
      properties:
        authKey:
          type: string
        email:
          type: string
        score:
          type: number
          format: double
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
    BaseJoin:
      allOf:
      - $ref: '#/components/schemas/TouchBasePostgresDomain'
      - type: object
        properties:
          baseId:
            type: string
          baseJoinAction:
            $ref: '#/components/schemas/BaseJoinAction'
    BaseJoinAction:
      type: string
      enum:
      - Request
      - Invite
    BaseJoinReq:
      type: object
      properties:
        baseId:
          type: string
        userId:
          type: string
        baseJoinAction:
          $ref: '#/components/schemas/BaseJoinAction'
    Success:
      type: object
      properties:
        status:
          type: string
    ListRes_Base_:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Base'
    Base:
      allOf:
      - $ref: '#/components/schemas/TouchBasePostgresDomain'
      - type: object
        properties:
          name:
            type: string
          score:
            type: number
            format: double
          imageUrl:
            type: string
          active:
            type: boolean
          members:
            type: array
            items:
              $ref: '#/components/schemas/BaseMember'
          chats:
            type: array
            items:
              $ref: '#/components/schemas/Chat'
    BaseMember:
      allOf:
      - $ref: '#/components/schemas/TouchBasePostgresDomain'
      - type: object
        properties:
          base:
            $ref: '#/components/schemas/Base'
          user:
            $ref: '#/components/schemas/User'
          role:
            $ref: '#/components/schemas/Role'
    Role:
      type: string
      enum:
      - MEMBER
      - ADMIN
    Chat:
      allOf:
      - $ref: '#/components/schemas/TouchBasePostgresDomain'
      - type: object
        properties:
          base:
            $ref: '#/components/schemas/Base'
          chatpiChatId:
            type: string
    BaseReq:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
    ListReq_MemberReq_:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/MemberReq'
    MemberReq:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/Role'
    ListReq_UUID_:
      type: object
      properties:
        list:
          type: array
          items:
            type: string
            format: uuid
    Mission:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    ProgressReq:
      type: object
      properties:
        value:
          type: integer
          format: int32